libhwc1: Fix large mmap() and munmap() length mismatches
The size of struct ExynosVideoMeta, used to determine lengths of
mappings in mmap() calls, is very large compared to the length used in
corresponding munmap() calls - 64 bytes. This causes quite some memory
leakage.

Under these BSP sources, ExynosVideoMeta has a size of 400 bytes. This
means that, on every time mmap() and munmap() are called with their
mismatched sizes in libhwc1, 400-64=336 bytes are undesirably leaked.
While being very problematic on its own, in theory for real usage, that
could eventually cause problems particularly after many days of usage
with very heavy video playback (possibly even encountered by someone out
there, but unreported or ignored as a 'one-off thing'; no problems were
known before with these BSP sources).

Fix this problem by matching lengths used for the mmap() and munmap()
calls.

Also, don't map the entire size of the ExynosVideoMeta struct - Since
only the 'eType' member in it is checked in between the mmap() and
munmap() calls, and it's right at the start of the struct, the size of
only the 'eType' member can be mapped (and unmapped), which is only a
tiny 4 bytes and seems to work fine.

For some trivia, this all looks like an oversight from someone at
Samsung. In some really old BSP sources [1], mmap(), as well as
munmap(), is called with a length of 64. When the mmap() calls were
changed for the lengths to be based on the size of ExynosVideoMeta, the
corresponding munmap() calls were missed and left at 64.

 1: https://github.com/LineageOS/android_hardware_samsung_slsi_exynos/blob/fb9857c5aa8923cc043b8f8a91bd24339c4bf528/libvppdisplay/ExynosDisplay.cpp#L1307

Change-Id: Ib3591a8e952e0e1dbf1ccfc9aec36d908ca1db29
Signed-off-by: Fakeman <55951275+fakemanoan@users.noreply.github.com>